Moved Param into Query. SPECIALIZEd Parametric refs.
- Moved Param module contents to Query module.
- Updated base to insert SPECIALIZE pragmas to ensure baseline performance in retained despite adding Parametric typeclass.
- Removed all references to Param module.
